add some list/dictionary/set comprehension samples, add datetime timezone sample

+++ sources/extras/logdb_sample.py
@@ -64,7 +64,7 @@ def adapt_datetime_iso(val: datetime.datetime) -> str :
     print(f"adapt_datetime_iso(val) : {val} : type(val) : {type(val)}")
     return val.isoformat()
 
-def adapt_datetime_epoch(val: datetime.datetime) -> float:
+def adapt_datetime_epoch(val: datetime.datetime) -> int:
     """Adapt datetime.datetime to Unix timestamp."""
     print(f"adapt_datetime_epoch(val) : {val} : type(val) : {type(val)}")
     return int(val.timestamp())
@@ -82,7 +82,7 @@ def convert_datetime(val: str) -> datetime.datetime:
     print(f"convert_datetime(val) : {val} : type(val) : {type(val)}")
     return datetime.datetime.fromisoformat(val.decode())
 
-def convert_timestamp(val: float) -> datetime.datetime:
+def convert_timestamp(val: int) -> datetime.datetime:
     """Convert Unix epoch timestamp to datetime.datetime object."""
     print(f"convert_timestamp(val) : {val} : type(val) : {type(val)}")
     return datetime.datetime.fromtimestamp(int(val))
